# fill faster when not doing aa