before:
```
name ............................... min time      avg time    std dv   runs
cairo1 ............................. 0.112 ms      0.114 ms    ±0.004  x1000
pixie1 ............................. 0.498 ms      0.504 ms    ±0.004  x1000
```

after:
```
name ............................... min time      avg time    std dv   runs
cairo1 ............................. 0.113 ms      0.115 ms    ±0.004  x1000
pixie1 ............................. 0.185 ms      0.189 ms    ±0.004  x1000
```